Map(Config)

+++ src/Map.cpp
@@ -30,8 +30,8 @@ Map::Map(std::string const & mapFile)
 
 Map::Map(Config::Param cfg)
 {
-	_width = cfg["width"];
-	_height = cfg["height"];
+	_width = static_cast<unsigned int>(cfg["width"]);
+	_height = static_cast<unsigned int>(cfg["height"]);
 	_actions[LEVEL_BOMB_EXPLODED] = &Map::bombExploded;
 	_actions[BLOCK_DESTROYED] = &Map::blockDestroyed;
